Alright, Protobuf it is.
I’ve given in to the madness that is protobuf, it consumes me now and
fills me with the hopefulness of not having to deal with JNI and
writing our own boilerplate, repeating serial code. For those
interested in using this: Run make in the protocol directory and it
should create the required .h and .c files for use on Arduino. We’ll
need to get a process set up for Java too but we’ve done it before so
it shouldn’t be too hard. Nanopb is a great library — really tiny
footprint and easy to use on Arduino. I’m moving to eclipse’s Arduino
plugin, I advise everyone else to do the same.